Defer creating shortcuts
We don't have to remove them when the MainActivityAlias is disabled.
