+++ app/src/main/kotlin/com/github/kr328/clash/app/MainApplication.kt
@@ -2,18 +2,10 @@ package com.github.kr328.clash.app
 
 import android.app.Application
 import android.content.Context
-import android.content.Intent
-import androidx.core.content.pm.ShortcutInfoCompat
-import androidx.core.content.pm.ShortcutManagerCompat
-import androidx.core.graphics.drawable.IconCompat
 import com.github.kr328.clash.common.Global
-import com.github.kr328.clash.common.constants.Intents
 import com.github.kr328.clash.common.log.Log
-import com.github.kr328.clash.common.util.unsafeLazy
 import com.github.kr328.clash.glue.remote.Remote
-import com.github.kr328.clash.glue.store.UiStore
 import com.github.kr328.clash.glue.util.clashDir
-import com.github.kr328.clash.glue.util.mainIntent
 import com.github.kr328.clash.service.util.sendServiceRecreated
 import java.io.File
 import org.koin.android.ext.koin.androidContext
@@ -21,7 +13,6 @@ import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
 
 class MainApplication : Application() {
-  private val uiStore by unsafeLazy { UiStore(this) }
 
   override fun attachBaseContext(base: Context?) {
     super.attachBaseContext(base)
@@ -41,7 +32,6 @@ class MainApplication : Application() {
 
     if (processName == packageName) {
       Remote.launch()
-      setupShortcuts()
     } else {
       sendServiceRecreated()
     }
@@ -53,49 +43,6 @@ class MainApplication : Application() {
       androidContext(this@MainApplication)
       modules(appModule)
     }
-  }
-
-  private fun setupShortcuts() {
-    if (uiStore.hideAppIcon) {
-      // Prevent launcher activity not found.
-      ShortcutManagerCompat.removeAllDynamicShortcuts(this)
-      return
-    }
-
-    val icon = IconCompat.createWithResource(this, R.mipmap.ic_launcher)
-    val flags =
-      Intent.FLAG_ACTIVITY_NEW_TASK or
-        Intent.FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS or
-        Intent.FLAG_ACTIVITY_NO_ANIMATION
-
-    val toggle =
-      ShortcutInfoCompat.Builder(this, "toggle_clash")
-        .setShortLabel(getString(R.string.shortcut_toggle_short))
-        .setLongLabel(getString(R.string.shortcut_toggle_long))
-        .setIcon(icon)
-        .setIntent(mainIntent(action = Intents.ACTION_TOGGLE_CLASH).addFlags(flags))
-        .setRank(0)
-        .build()
-
-    val start =
-      ShortcutInfoCompat.Builder(this, "start_clash")
-        .setShortLabel(getString(R.string.shortcut_start_short))
-        .setLongLabel(getString(R.string.shortcut_start_long))
-        .setIcon(icon)
-        .setIntent(mainIntent(action = Intents.ACTION_START_CLASH).addFlags(flags))
-        .setRank(1)
-        .build()
-
-    val stop =
-      ShortcutInfoCompat.Builder(this, "stop_clash")
-        .setShortLabel(getString(R.string.shortcut_stop_short))
-        .setLongLabel(getString(R.string.shortcut_stop_long))
-        .setIcon(icon)
-        .setIntent(mainIntent(action = Intents.ACTION_STOP_CLASH).addFlags(flags))
-        .setRank(2)
-        .build()
-
-    ShortcutManagerCompat.setDynamicShortcuts(this, listOf(toggle, start, stop))
   }
 
   private fun extractGeoFiles() {
